closes #126 Introduce COBOL Sections in the transformed code
isParagraphHeader now takes three parameters. Instead of looking through each character of the source line, it checks if there is only one token, and if that token is followed by a period on the same or the next line.

Tests has been added, in order to test more formats for paragraph headers.

Signed-off-by: David Kaan <dak@bankdata.dk>